Reporting functions for repMean, repTable, repQuantile, and repGlm

Summarizes the output of the four main functions repMean, repTable, repQuantile, and repGlm, and provides a single data.frame with all results.

Usage

report(repFunOut, trendDiffs = deprecated(), add=list(),
       exclude = c("NcasesValid", "var"),
       printGlm = FALSE, round = TRUE, digits = 3, printDeviance = FALSE,
       printSE_correction = FALSE)
report2(repFunOut, add=list(), exclude = c("NcasesValid", "var"), printGlm = FALSE,
       round = TRUE, digits = 3, printDeviance = FALSE, printSE_correction = FALSE)

Arguments

repFunOut

output of one of the four eatRep-functions.

trendDiffs

deprecated. In earlier versions of the package, this argument was used to determine differences in trends. As differences in trends are equivalent to the trend of differences (no matter whether group or cross-level differences), the argument was deprecated. If the user specifies group or cross-level difference along with trends, trends of differences are computed as well.

add

Optional: additional columns for output. See examples of the jk2-functions

exclude

Which parameters should be excluded from reporting?

printGlm

Only relevant for repGlm: print summary on console?

round

Logical: should the results be rounded to a limited number of digits?

digits

How many digits should be used for rounding?

printDeviance

Only relevant for repGlm when other than the identical function is used as link function, and if printGlm is TRUE. Should the deviance information printed additionally? Note: To print deviance information, the argument poolMethod of the repGlm function must be set to "scalar".

printSE_correction

Logical: Print the differences of original SEs of cross differences (method "old") and SEs obtained by the "wec" or "rep" method.

Value

report and report2 differ in the output which is returned. The output of report2 is optimized for further processing, i.e. drawing plots by means of the eatPlot package. For report, the output is a data frame with at least nine columns.

group

Denotes the group an analysis belongs to. If no groups were specified and/or analysis for the whole sample were requested, the value of ‘group’ is ‘wholeGroup’.

depVar

Denotes the name of the dependent variable in the analysis.

modus

Denotes the mode of the analysis. For example, if a JK2 regression analysis was conducted, ‘modus’ takes the value ‘JK2.glm’. If a mean analysis without any replicates was conducted, ‘modus’ takes the value ‘CONV.mean’.

comparison

Denotes whether group mean comparisons or cross-level comparisons were conducted. Without any comparisons, ‘comparison’ takes the value ‘NA’

parameter

Denotes the parameter of the corresponding analysis. If regression analysis was applied, the regression parameter is given. Amongst others, the ‘parameter’ column takes the values ‘(Intercept)’ and ‘gendermale’ if ‘gender’ was the independent variable, for instance. If mean analysis was applied, the ‘parameter’ column takes the values ‘mean’, ‘sd’, ‘var’, or ‘Nvalid’. See the examples of repMean,repTable, repQuantile, or repGlm for further details.

depVar

Denotes the name of the dependent variable (only if repGlm was called before)

est

Denotes the estimate of the corresponding analysis.

se

Denotes the standard error of the corresponding estimate.

p

Denotes the p value of the estimate.

For report2, the output is a list with four data.frames. The first data.frame plain summarizes the results in human-readable form. The data.frames 2 to 4 (comparisons, group, estimate) are redundant to plain and contain the results in a technical presentation suitable for further processing in eatPlot.

plain

The complete results in human-readable form.

comparison

An allocation table that indicates which comparison (group comparison or cross-level comparison) relates to which groups.

group

A table that assigns an ID to each analysis unit. This makes it easier later on to read from the output which comparison relates to which groups. This simplifies the assignment, especially when comparing comparisons (i.e., cross-level differences of group differences).

estimate

The results of the analyses, assigned to their IDs.

Author

Benjamin Becker, Sebastian Weirich

Examples

### see examples of the eatRep main functions.